requirements.txt
# Core web framework
flask==3.0.3

# MySQL connector (matches your current version)
mysql-connector-python==9.0.0

# Environment variable management
python-dotenv==1.0.1

# Production WSGI server (strongly recommended for Render)
gunicorn==23.0.0